+++ webpack.config.js
@@ -1,0 +1,4 @@
+let webpack = require('vortex-api/bin/webpack').default;
+
+module.exports = webpack('vortex-modlist-backup', __dirname, 4);
+