+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Changelog
 
+#### Version 0.2.0
+ * Moved to Typescript
+ * Changed error popups to success notifications
+ * Backuping up modlist no longer removes mods from games that aren't part of Vortex anymore
+
 #### Version 0.1.1
  * Added a file selector to choose where to back up to and restore from